pyproject.toml
[project]
name = "kagi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "pyjwt>=2.10.1",
    "typer>=0.15",
    "rich>=13.0",
    "pygments>=2.19.2",
]

[project.scripts]
kagi = "kagi_client.cli:app"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "respx>=0.22",
    "ruff>=0.9",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["kagi_client"]

[tool.ruff]
line-length = 88
